+++ R/outline.R
@@ -90,7 +90,7 @@ file_outline <- function(pattern = NULL,
                          recent_only = FALSE) {
   # To contribute to this function, take a look at .github/CONTRIBUTING
 
-  if (length(path) == 1L && interactive() && rstudioapi::isAvailable()) {
+  if (length(path) == 1L && rlang::is_interactive() && is_rstudio()) {
     is_active_doc <- identical(path, active_rs_doc())
   } else {
     is_active_doc <- FALSE
@@ -134,8 +134,8 @@ file_outline <- function(pattern = NULL,
     file_content <- dplyr::bind_rows(file_content, .id = "file")
   }
 
-  suppressMessages(
-    in_active_project <- tryCatch(
+  in_active_project <- suppressMessages(
+    tryCatch(
       identical(suppressWarnings(proj_get2()), dir_common),
       error = function(e) FALSE
     )
@@ -197,7 +197,7 @@ file_outline <- function(pattern = NULL,
   }
   # File outline ===================
   # strip outline element .data$outline = `# Section 1` becomes `Section 1`
-  file_sections1 <- display_outline_element(file_sections0)
+  file_sections1 <- display_outline_element(file_sections0, dir_common)
 
   # Create hyperlink in console
   file_sections <- construct_outline_link(file_sections1, is_saved_doc, dir_common, pattern)
@@ -215,7 +215,7 @@ file_outline <- function(pattern = NULL,
     )
   file_sections$recent_only <- recent_only
 
-  if (any(duplicated(file_sections$outline_el))) {
+  if (anyDuplicated(file_sections$outline_el) > 0L) {
     file_sections <- scrub_duplicate_outline(file_sections)
   }
   file_sections <- dplyr::relocate(
@@ -251,14 +251,14 @@ proj_outline <- function(pattern = NULL, proj = proj_get2(), work_only = TRUE, d
     ))
   }
 
-  if (!fs::dir_exists(proj)) { # when referring to a project by name.
-    proj_dir <- proj_list(proj)
-  } else {
+  if (fs::dir_exists(proj)) {
     if (!is_active_proj) {
       cli::cli_warn("Use {.fn dir_outline} for that.")
     }
 
     proj_dir <- proj
+  } else { # when referring to a project by name.
+    proj_dir <- proj_list(proj)
   }
 
   if (!rlang::has_length(proj_dir, 1)) {
@@ -310,11 +310,7 @@ dir_outline <- function(pattern = NULL, path = ".", work_only = TRUE, dir_tree =
   if (recurse && !identical(Sys.getenv("TESTTHAT"), "true")) {
     # Remove examples from outline and test example files to avoid clutter
     # examples don't help understand a project.
-    file_list_to_outline <- fs::path_filter(
-      file_list_to_outline,
-      regexp = "testthat/_ref/|example-file",
-      invert = TRUE
-    )
+    file_list_to_outline <- exclude_example_files(file_list_to_outline)
   }
 
   if (any(grepl("README.Rmd", file_list_to_outline))) {
@@ -325,7 +321,7 @@ dir_outline <- function(pattern = NULL, path = ".", work_only = TRUE, dir_tree =
 
     fs::dir_tree(
       path = dir,
-      regexp = "R/.+|qmd|Rmd|_files|~\\$|*.Rd|_snaps|testthat.R|Rmarkdown|docs/",
+      regexp = "R/.+|qmd|Rmd|_files|~\\$|*.Rd|_snaps|tests/testthat.R|Rmarkdown|docs/",
       recurse = recurse,
       invert = TRUE
     )
@@ -333,6 +329,28 @@ dir_outline <- function(pattern = NULL, path = ".", work_only = TRUE, dir_tree =
   file_outline(path = file_list_to_outline, pattern = pattern, work_only = work_only, dir_common = dir, alpha = alpha, recent_only = recent_only)
 }
 
+exclude_example_files <- function(path) {
+  # styler tests examples may not work..
+
+  regexp_exclude <- paste(
+    "vignettes/test/", # test vignettes
+    "LICENSE.md", # avoid indexing this.
+    "tests/(performance-monitor|gt-examples/|testthat/scope-|testthat/assets|testthat/_outline|testthat/testTestWithFailure|testthat/testTest/|testthat/test-parallel/|testthat/test-list-reporter/)", # example files in usethis, pkgdown, reuseme, devtools, etc.
+    "inst/((rmarkdown/)?templates/|example-file/|examples/rmd/|tutorials/)", # license templates in usethis
+    "revdep/", # likely don't need to outline revdep/, use dir_outline() to find something in revdep/
+    "themes/hugo-theme-console/", # protect blogdown
+    "vignettes/.+\\.R$", # generated files
+    "RcppExports.R",
+    "pkgdown/assets",
+    sep = "|"
+  )
+
+  fs::path_filter(
+    path,
+    regexp = regexp_exclude,
+    invert = TRUE
+  )
+}
 # Print method -------------------
 
 #' @export
@@ -419,7 +437,16 @@ print.outline_report <- function(x, ...) {
       base_name <- c(base_name, " ", title_el)
     }
 
-    cli::cli_h3(base_name)
+    # TRICK need tryCatch when doing something, withCallingHandlers when only rethrowing?
+    tryCatch(
+      cli::cli_h3(base_name),
+      error = function(e) {
+        # browser()
+        cli::cli_h3(escape_markup(base_name))
+        # print(base_name)
+        # rlang::abort("Could not parse by cli", parent = e)
+      }
+    )
 
     if (recent_only) {
       if (i %in% is_recently_modified) {
@@ -465,7 +492,7 @@ keep_outline_element <- function(.data) {
       # What to keep in .R files
       (!is_md & is_section_title_source) |
       # What to keep anywhere
-       is_tab_or_plot_title | is_todo_fixme | is_test_name | is_cross_ref | is_function_def # | is_cli_info # TODO reanable cli info
+      is_tab_or_plot_title | is_todo_fixme | is_test_name | is_cross_ref | is_function_def # | is_cli_info # TODO reanable cli info
   )
   dat$simplify_news <- NULL
   dat
@@ -475,23 +502,28 @@ keep_outline_element <- function(.data) {
 # Includes removing headings comments
 # Remove title =
 # Removing quotes, etc.
-display_outline_element <- function(.data) {
+display_outline_element <- function(.data, dir_common) {
   x <- .data
-  x$outline_el <- purrr::map_chr(x$content, link_gh_issue) # to add link to GitHub.
+  org_repo <- find_pkg_org_repo(dir_common, unique(x$file))
+  if (!is.null(org_repo)) {
+    x$outline_el <- link_local_gh_issue(x$content, org_repo)
+  } else {
+    x$outline_el <- x$content
+  }
+  x$outline_el <- purrr::map_chr(x$outline_el, \(x) link_gh_issue(x, org_repo)) # to add link to GitHub.
   x$outline_el <- purrr::map_chr(x$outline_el, markup_href)
   x <- dplyr::mutate(
     x,
     outline_el = dplyr::case_when(
-      is_todo_fixme ~ stringr::str_extract(outline_el, "(TODO.+)|(FIXME.+)|(WORK.+)"),
-      is_test_name ~ stringr::str_extract(outline_el, "test_that\\(['\"](.+)['\"]", group = 1),
+      is_todo_fixme ~ stringr::str_extract(outline_el, "(TODO.+)|(FIXME.+)|(WORK.+)|(BOOK.+)"),
+      is_test_name ~ stringr::str_extract(outline_el, "test_that\\(['\"](.+)['\"],\\s?\\{", group = 1),
       is_cli_info ~ stringr::str_extract(outline_el, "[\"'](.{5,})[\"']") |> stringr::str_remove_all("\""),
-      is_tab_or_plot_title ~ stringr::str_extract(outline_el, "title = [\"']([^\"]{5,})[\"']", group = 1),
-      is_chunk_cap_next & !is_chunk_cap ~ stringr::str_remove_all(outline_el, "\\s?\\#\\|\\s+"),
+      is_tab_or_plot_title ~ stringr::str_extract(outline_el, "title =[^\"']*[\"']([^\"]{5,})[\"']", group = 1), is_chunk_cap_next & !is_chunk_cap ~ stringr::str_remove_all(outline_el, "\\s?\\#\\|\\s+"),
       is_chunk_cap ~ stringr::str_remove_all(stringr::str_extract(outline_el, "(cap|title)\\:\\s*(.+)", group = 2), "\"|'"),
       is_cross_ref ~ stringr::str_remove_all(outline_el, "^(i.stat\\:\\:)?.cdocs_lin.s\\(|[\"']\\)$|\""),
       is_doc_title ~ stringr::str_remove_all(outline_el, "subtitle\\:\\s?|title\\:\\s?|\"|\\#\\|\\s?"),
       is_section_title & !is_md ~ stringr::str_remove(outline_el, "^\\s{0,4}\\#+\\s+|^\\#'\\s\\#+\\s+"), # Keep inline markup
-      is_section_title & is_md ~ stringr::str_remove_all(outline_el, "^\\#+\\s+|\\{.+\\}"), # strip cross-refs.
+      is_section_title & is_md ~ stringr::str_remove_all(outline_el, "^\\#+\\s+|\\{.+\\}|<(a href|img src).+$"), # strip cross-refs.
       is_function_def ~ stringr::str_extract(outline_el, "(.+)\\<-", group = 1) |> stringr::str_trim(),
       .default = stringr::str_remove_all(outline_el, "^\\s*\\#+\\|?\\s?(label:\\s)?|\\s?[-\\=]{4,}")
     ),
@@ -570,13 +602,13 @@ define_important_element <- function(.data) {
 }
 
 construct_outline_link <- function(.data, is_saved_doc, dir_common, pattern) {
-  rs_avail_file_link <- rstudioapi::isAvailable("2023.09.0.375") # better handling after
+  rs_avail_file_link <- is_rstudio("2023.09.0.375") # better handling after
   .data <- define_important_element(.data)
 
   if (is.null(dir_common) || !nzchar(dir_common)) {
     dir_common <- "Don't remove anything if not null"
   }
-  .data$rs_version <- ifelse(!rstudioapi::isAvailable("2023.12.0.274") && rstudioapi::isAvailable(), ".", "")
+  .data$rs_version <- ifelse(!is_rstudio("2023.12.0.274") && is_rstudio(), ".", "")
   .data$has_inline_markup <- dplyr::coalesce(stringr::str_detect(.data$outline_el, "\\{|\\}"), FALSE)
   .data$is_saved_doc <- is_saved_doc
   .data <- dplyr::mutate(
@@ -625,20 +657,17 @@ construct_outline_link <- function(.data, is_saved_doc, dir_common, pattern) {
     outline_el2 = dplyr::coalesce(outline_el2, outline_el)
   )
 
-  .data <- dplyr::mutate(.data,
-    link = paste0(outline_el2, " {.path ", file, ":", line, "}"),
-    # rstudioapi::documentOpen works in the visual mode!! but not fully.
-    file_path = .data$file,
-    is_saved_doc = .env$is_saved_doc,
-
-    # May have caused CI failure
-    text_in_link = stringr::str_remove(file_path, as.character(.env$dir_common)) |> stringr::str_remove("^/"),
-    # decide which is important
-    style_fun = dplyr::case_match(importance,
-      "not_important" ~ "cli::style_italic('i')", # cli::style_inverse for bullets
-      "important" ~ "cli::style_inverse('i')",
-      .default = NA
-    )
+  .data$link <- paste0(.data$outline_el2, " {.path ", .data$file, ":", .data$line, "}")
+  # rstudioapi::documentOpen works in the visual mode!! but not fully.
+  .data$file_path <- .data$file
+  .data$is_saved_doc <- is_saved_doc
+  # May have caused CI failure
+  .data$text_in_link <- sub(as.character(dir_common), "", .data$file_path)
+  .data$text_in_link <- sub("^/", "", .data$text_in_link)
+  .data$style_fun <- dplyr::case_match(.data$importance,
+    "not_important" ~ "cli::style_italic('i')", # cli::style_inverse for bullets
+    "important" ~ "cli::style_inverse('i')",
+    .default = NA_character_
   )
 
   if (anyNA(.data$style_fun)) {
@@ -664,7 +693,17 @@ construct_outline_link <- function(.data, is_saved_doc, dir_common, pattern) {
     rs_version = NULL,
     outline_el2 = NULL,
     condition_to_truncate = NULL,
-    condition_to_truncate2 = NULL
+    condition_to_truncate2 = NULL,
+    style_fun = NULL,
+    is_saved_doc = NULL,
+    is_roxygen_comment = NULL,
+    is_news = NULL,
+    # I may put it pack
+    importance = NULL,
+    # may be useful for debugging
+    before_and_after_empty = NULL,
+    # may be useful for debugging
+    has_inline_markup = NULL
   ) |>
     dplyr::filter(is.na(outline_el) | grepl(pattern, outline_el, ignore.case = TRUE))
 }