+++ tests/testthat/test-rename.R
@@ -1,5 +1,5 @@
 describe("rename_files2()", {
-  og_file <- fs::path_real(test_path("_ref", "my-analysis.R"))
+  og_file <- fs::path_real(test_path("_outline", "my-analysis.R"))
   # temp dir + change working directory
   tmp_dir <- withr::local_tempdir()
   fs::dir_create(
@@ -123,7 +123,7 @@ test_that("file testing are working as expected", {
 })
 
 test_that("force and action are deprecated", {
-  file <- withr::local_tempfile(fileext = ".R", lines = c("# x1"))
+  file <- withr::local_tempfile(fileext = ".R", lines = "# x1")
   file2 <- withr::local_tempfile(fileext = ".R")
   unlink(file2)
   lifecycle::expect_deprecated(